django_reference_pipeline/requirements.txt
# Django API
Django>=5.0,<6
djangorestframework>=3.15
django-cors-headers>=4.3
django-environ>=0.11

# Async queue
celery[redis]>=5.3
redis>=5.0

# S3
boto3>=1.34
django-storages[boto3]>=1.14

# Files / images
Pillow>=10.0
